pyproject.toml
[tool.poetry]
name = "med-enterprise-dash"
version = "0.1.0"
description = "A tool for managing patient data."
authors = ["vaeringjar <vaeringjar@peers.community>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
pyramid = "^1.10.4"
pyramid_jinja2 = "^2.8"
python_cas = "^1.5.0"
toml = "^0.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
pyclean = "^2.0.0"
mypy = "^0.790"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
